Combine all the code into three files index main and style
Combine all JavaScript code into `main.js` and remove the `Scripts` directory.

* Merge the content of `Scripts/game class and values/game class.js`, `Scripts/game class and values/game varables.js`, `Scripts/gamefuntion.js`, and `Scripts/unity.js` into `main.js`.
* Delete `Scripts/game class and values/game class.js`, `Scripts/game class and values/game varables.js`, `Scripts/gamefuntion.js`, and `Scripts/unity.js`.
* Remove the `Scripts` directory.
* Add a new `items.json` file to store items with an initial structure.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Vortextbloons/RPG?shareId=XXXX-XXXX-XXXX-XXXX).
